+++ src/analysis/const_fold.c
@@ -18,7 +18,7 @@ int eval_const_int_expr(ASTNode *node, ParserContext *ctx, long long *out_val)
     case NODE_EXPR_LITERAL:
         if (node->literal.type_kind == LITERAL_INT)
         {
-            *out_val = node->literal.int_val;
+            *out_val = (long long int)(node->literal.int_val);
             return 1;
         }
         return 0;
